Move image details to test details modal (#401)
